requirements.txt
numpy
opencv-python
Pillow
streamlit
streamlit-option-menu
plotly
pandas
seaborn